refactor: move duties to separate modules
