+++ src/duty/attestation/index.ts
@@ -1,0 +1,2 @@
+export * from './attestation.module';
+export * from './attestation.service';
