+++ src/duty/index.ts
@@ -1,0 +1,2 @@
+export * from './duty.module';
+export * from './duty.service';
