+++ src/duty/propose/index.ts
@@ -1,0 +1,2 @@
+export * from './propose.module';
+export * from './propose.service';
