+++ src/duty/propose/propose.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+
+import { ConsensusProviderModule } from 'common/eth-providers';
+
+import { ProposeService } from './propose.service';
+
+@Module({
+  imports: [ConsensusProviderModule],
+  providers: [ProposeService],
+  exports: [ProposeService],
+})
+export class ProposeModule {}
