+++ src/duty/state/state.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+
+import { ConsensusProviderModule } from 'common/eth-providers';
+
+import { StateService } from './state.service';
+
+@Module({
+  imports: [ConsensusProviderModule],
+  providers: [StateService],
+  exports: [StateService],
+})
+export class StateModule {}
