+++ src/duty/sync/index.ts
@@ -1,0 +1,2 @@
+export * from './sync.module';
+export * from './sync.service';
