+++ src/duty/sync/sync.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+
+import { ConsensusProviderModule } from 'common/eth-providers';
+
+import { SyncService } from './sync.service';
+
+@Module({
+  imports: [ConsensusProviderModule],
+  providers: [SyncService],
+  exports: [SyncService],
+})
+export class SyncModule {}
